requirements.txt
numpy
streamlit
streamlit-chat==0.0.2.1
torch
torchvision
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
Cython==0.29.23
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
nltk==3.6.6
numpy==1.21.0
Pillow==9.0.1
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2021.8.3
requests==2.25.1
sentencepiece==0.1.95
six==1.16.0
#torch==1.4.0
torchtext==0.5.0
#torchvision==0.5.0
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
